src/requirements.txt
asgiref==3.4.1
Django==3.2.5
gunicorn==20.1.0
mysqlclient==2.0.3
PyMySQL==1.0.2
python-dotenv==0.18.0
pytz==2021.1
sqlparse==0.4.1
django-ninja